pyproject.toml
[project]
name = "kubani"
version = "0.1.0"
description = "Kubernetes cluster automation and management for Tailscale networks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "textual>=0.47.0",
    "pydantic>=2.5.0",
    "ruamel.yaml>=0.18.0",
    "kubernetes>=29.0.0",
    "ansible-runner>=2.3.0",
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
    "pytest-cov>=4.1.0",
    "tomli>=2.0.0",
    "ruff>=0.14.7",
    "ty>=0.0.1a30",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
cluster-mgr = "cluster_manager.cli:app"
cluster-tui = "cluster_manager.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cluster_manager"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=cluster_manager",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--hypothesis-show-statistics",
]

[tool.coverage.run]
source = ["cluster_manager"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.hypothesis]
max_examples = 100
deadline = 5000
derandomize = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Ignore line length errors

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
